requirements_short.txt
torchcrepe
praat-parselmouth==0.4.1
scikit-image
ipython
ipykernel
pyloudnorm
webrtcvad
h5py
einops
pycwt
torchmetrics==0.5
pytorch_lightning==1.3.3